+++ packages/widgetbook/test/generator/doc_comment/doc_comment.stories.g.dart
@@ -6,16 +6,15 @@ part of 'doc_comment.stories.dart';
 // StoryGenerator
 // **************************************************************************
 
-typedef _Component = Component<DocCommentWidget, DocCommentWidgetArgs>;
+typedef _Component = Component<DocCommentWidget, StoryArgs<DocCommentWidget>>;
 typedef _Scenario = DocCommentWidgetScenario;
 typedef _Defaults = DocCommentWidgetDefaults;
 typedef _Story = DocCommentWidgetStory;
 typedef _Args = DocCommentWidgetArgs;
 final DocCommentWidgetComponent =
-    Component<DocCommentWidget, DocCommentWidgetArgs>(
-      name: meta.name ?? 'DocCommentWidget',
-      path: meta.path ?? '',
-      docsBuilder: meta.docsBuilder,
+    Component<DocCommentWidget, StoryArgs<DocCommentWidget>>(
+      name: 'DocCommentWidget',
+      path: '',
       docComment: r'''A widget with documentation.
 
 This has an empty line above.''',