+++ packages/widgetbook/test/generator/doc_comment_code/doc_comment_code.stories.g.dart
@@ -6,16 +6,16 @@ part of 'doc_comment_code.stories.dart';
 // StoryGenerator
 // **************************************************************************
 
-typedef _Component = Component<DocCommentCodeWidget, DocCommentCodeWidgetArgs>;
+typedef _Component =
+    Component<DocCommentCodeWidget, StoryArgs<DocCommentCodeWidget>>;
 typedef _Scenario = DocCommentCodeWidgetScenario;
 typedef _Defaults = DocCommentCodeWidgetDefaults;
 typedef _Story = DocCommentCodeWidgetStory;
 typedef _Args = DocCommentCodeWidgetArgs;
 final DocCommentCodeWidgetComponent =
-    Component<DocCommentCodeWidget, DocCommentCodeWidgetArgs>(
-      name: meta.name ?? 'DocCommentCodeWidget',
-      path: meta.path ?? '',
-      docsBuilder: meta.docsBuilder,
+    Component<DocCommentCodeWidget, StoryArgs<DocCommentCodeWidget>>(
+      name: 'DocCommentCodeWidget',
+      path: '',
       docComment: r'''Displays a [label].
 
 ```dart