+++ sandbox/lib/[sam]/types_table.stories.dart
@@ -5,7 +5,7 @@ import 'types_table.dart';
 
 part 'types_table.stories.g.dart';
 
-final meta = Meta<TypesTable>(
+final component = ComponentMeta(
   docsBuilder: (blocks) => blocks.replaceFirst<DartCommentDocBlock>(
     const TextDocBlock('''
 1. All possible combinations of parameters (required, optional, with default values)
@@ -14,6 +14,8 @@ final meta = Meta<TypesTable>(
 '''),
   ),
 );
+
+const meta = Meta(TypesTable.new);
 
 final $Default = _Story(
   setup: (context, child, args) {