+++ sandbox/lib/[sam]/types_table.stories.g.dart
@@ -11,15 +11,15 @@ part of 'types_table.stories.dart';
 // StoryGenerator
 // **************************************************************************
 
-typedef _Component = Component<TypesTable, TypesTableArgs>;
+typedef _Component = Component<TypesTable, StoryArgs<TypesTable>>;
 typedef _Scenario = TypesTableScenario;
 typedef _Defaults = TypesTableDefaults;
 typedef _Story = TypesTableStory;
 typedef _Args = TypesTableArgs;
-final TypesTableComponent = Component<TypesTable, TypesTableArgs>(
-  name: meta.name ?? 'TypesTable',
-  path: meta.path ?? '[sam]',
-  docsBuilder: meta.docsBuilder,
+final TypesTableComponent = Component<TypesTable, StoryArgs<TypesTable>>(
+  name: component.name ?? 'TypesTable',
+  path: component.path ?? '[sam]',
+  docsBuilder: component.docsBuilder,
   docComment: null,
   stories: [$Default..$generatedName = 'Default'],
 );